split built-in artifact document helpers
