+++ .sampo/changesets/issue-377-built-in-artifact-document-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split built-in block artifact document emitters and attribute spec tables into focused runtime helper modules while preserving generated output.
